requirements.txt
redis==3.5.3
emojis==0.6.0
celery==5.0.1
requests==2.20.0
psycopg2-binary==2.8.6
pytelegrambotapi==3.7.3
